purge circuit/gates folder
