--- kimchi/src/circuits/gates/mod.rs
@@ -1,3 +1,0 @@
-pub mod endosclmul;
-pub mod poseidon;
-pub mod varbasemul;
