+++ kimchi/src/circuits/mod.rs
@@ -6,7 +6,6 @@ pub mod constraints;
 pub mod domains;
 pub mod expr;
 pub mod gate;
-pub mod gates;
 pub mod polynomial;
 pub mod polynomials;
 pub mod scalars;
