+++ kimchi/src/tests/poseidon.rs
@@ -1,17 +1,12 @@
-use crate::{
-    circuits::{
-        gate::CircuitGate,
-        gates::poseidon::{self, ROUNDS_PER_ROW},
-        wires::{Wire, COLUMNS},
-    },
-    prover_index::testing::new_index_for_test,
-    verifier::batch_verify,
-};
-use crate::{prover::ProverProof, prover_index::ProverIndex};
+use std::time::Instant;
+use std::{io, io::Write};
+
 use ark_ff::{UniformRand, Zero};
 use ark_poly::{univariate::DensePolynomial, UVPolynomial};
 use array_init::array_init;
 use colored::Colorize;
+use rand::{rngs::StdRng, SeedableRng};
+
 use commitment_dlog::commitment::{b_poly_coefficients, ceil_log2, CommitmentCurve};
 use groupmap::GroupMap;
 use mina_curves::pasta::{
@@ -22,9 +17,18 @@ use oracle::{
     poseidon::{PlonkSpongeConstantsKimchi, SpongeConstants},
     sponge::{DefaultFqSponge, DefaultFrSponge},
 };
-use rand::{rngs::StdRng, SeedableRng};
-use std::time::Instant;
-use std::{io, io::Write};
+
+use crate::circuits::polynomials;
+use crate::circuits::polynomials::poseidon::ROUNDS_PER_ROW;
+use crate::{
+    circuits::{
+        gate::CircuitGate,
+        wires::{Wire, COLUMNS},
+    },
+    prover_index::testing::new_index_for_test,
+    verifier::batch_verify,
+};
+use crate::{prover::ProverProof, prover_index::ProverIndex};
 
 // aliases
 
@@ -120,7 +124,7 @@ fn positive(index: &ProverIndex<Affine>) {
             // index
             let first_row = h * (POS_ROWS_PER_HASH + 1);
 
-            poseidon::generate_witness(
+            polynomials::poseidon::generate_witness(
                 first_row,
                 oracle::pasta::fp_kimchi::params(),
                 &mut witness_cols,
